improving the courses page
Fixing this and stuff

Warning: The document is not mappable to XML 1.0 due to two consecutive hyphens in a comment.

At line 59, column 11

ody>↩    <!-------------------

Error: Bad value 40px for attribute height on element img: Expected a digit but saw p instead.

From line 63, column 17; to line 63, column 95

          <img src="assets/LOGO/LOGO_Do.png" alt="DO-BU Logo" height="40px" width="86px">↩

Error: Bad value 86px for attribute width on element img: Expected a digit but saw p instead.

From line 63, column 17; to line 63, column 95

          <img src="assets/LOGO/LOGO_Do.png" alt="DO-BU Logo" height="40px" width="86px">↩

Error: Element a not allowed as child of element ul in this context. (Suppressing further errors from this subtree.)

From line 75, column 21; to line 75, column 81

          <a href="sign-up.html" class="btn btn-light join-signup-btn">Join s

Contexts in which element a may be used:
Where phrasing content is expected.
Content model for element ul:
Zero or more li and script-supporting elements.
Warning: The document is not mappable to XML 1.0 due to two consecutive hyphens in a comment.

At line 81, column 11

av>↩↩    <!-------------------

Warning: The document is not mappable to XML 1.0 due to two consecutive hyphens in a comment.

At line 81, column 54

Courses Grid -----------------

Warning: The document is not mappable to XML 1.0 due to two consecutive hyphens in a comment.

At line 85, column 15

↩        <!-------------------

Error: Bad value assets/courses/Muay Thai.jpg for attribute src on element img: Illegal character in path segment: space is not allowed.

From line 123, column 21; to line 123, column 94

          <img src="assets/courses/Muay Thai.jpg" alt="Muay Thai" class="img-fluid">↩

Warning: The document is not mappable to XML 1.0 due to two consecutive hyphens in a comment.

At line 154, column 15

↩        <!-------------------

Warning: The document is not mappable to XML 1.0 due to two consecutive hyphens in a comment.

At line 168, column 11

iv>↩↩    <!-------------------

Warning: The document is not mappable to XML 1.0 due to two consecutive hyphens in a comment.

At line 197, column 11

er>↩↩    <!-------------------